requirements.txt
fastapi
uvicorn[standard]
sentence-transformers
faiss-cpu
yandex-cloud-ml-sdk
langchain